Created new package for gateways configurations
